refactor: extract the form validation and precognition machinery
Pull FormDefinitions single-pass validate() into a reusable FieldValidator,
and FormControllers precognition handling into a HandlesPrecognition trait,
so other endpoints can reuse both. No behavior change.
